+++ .changeset/stupid-flowers-push.md
@@ -1,0 +1,7 @@
+---
+'myst-frontmatter': patch
+'myst-config': patch
+'myst-cli': patch
+---
+
+Support parts in site config
