+++ packages/myst-frontmatter/src/site/validators.ts
@@ -4,6 +4,7 @@ import {
   incrementOptions,
   validateList,
   validateObject,
+  validateObjectKeys,
   validateString,
   validationError,
 } from 'simple-validators';
@@ -14,7 +15,7 @@ import type { ReferenceStash } from '../utils/referenceStash.js';
 import { validateAndStashObject } from '../utils/referenceStash.js';
 import { validateGithubUrl } from '../utils/validators.js';
 import { validateVenue } from '../venues/validators.js';
-import type { SiteFrontmatter } from './types.js';
+import { FRONTMATTER_ALIASES, PAGE_KNOWN_PARTS, type SiteFrontmatter } from './types.js';
 import { RESERVED_EXPORT_KEYS } from '../exports/validators.js';
 
 export function validateSiteFrontmatterKeys(value: Record<string, any>, opts: ValidationOptions) {
@@ -169,6 +170,40 @@ export function validateSiteFrontmatterKeys(value: Record<string, any>, opts: Va
           (output.options ??= {})[key] = val;
         }
       });
+    }
+  }
+  const partsOptions = incrementOptions('parts', opts);
+  let parts: Record<string, any> | undefined;
+  if (defined(value.parts)) {
+    parts = validateObjectKeys(
+      value.parts,
+      { optional: PAGE_KNOWN_PARTS, alias: FRONTMATTER_ALIASES },
+      { keepExtraKeys: true, suppressWarnings: true, ...partsOptions },
+    );
+  }
+  PAGE_KNOWN_PARTS.forEach((partKey) => {
+    if (defined(value[partKey])) {
+      parts ??= {};
+      if (parts[partKey]) {
+        validationError(`duplicate value for part ${partKey}`, partsOptions);
+      } else {
+        parts[partKey] = value[partKey];
+      }
+    }
+  });
+  if (parts) {
+    const partsEntries = Object.entries(parts)
+      .map(([k, v]) => {
+        return [
+          k,
+          validateList(v, { coerce: true, ...incrementOptions(k, partsOptions) }, (item, index) => {
+            return validateString(item, incrementOptions(`${k}.${index}`, partsOptions));
+          }),
+        ];
+      })
+      .filter((entry): entry is [string, string[]] => !!entry[1]?.length);
+    if (partsEntries.length > 0) {
+      output.parts = Object.fromEntries(partsEntries);
     }
   }
